requirements.txt
attrs==24.2.0
aws-cdk-lib==2.159.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==36.3.0
aws-cdk.integ-tests-alpha==2.154.1a0
aws-solutions-constructs.aws-cloudfront-apigateway==2.70.0
aws-solutions-constructs.aws-cloudfront-apigateway-lambda==2.70.0
aws-solutions-constructs.core==2.70.0
cattrs==23.2.3
certifi==2024.8.30
charset-normalizer==3.3.2
constructs==10.3.0
DateTime==5.5
idna==3.10
importlib_resources==6.4.5
jsii==1.103.1
publication==0.0.3
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.2.3
zope.interface==7.0.3
